// client/lib/api/models/Axis.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Axis = {
    id?: number;
    zoneId?: number;
    key?: Axis.key;
    label?: string;
};
export namespace Axis {
    export enum key {
        NORD = 'nord',
        EST = 'est',
        SUD = 'sud',
        OUEST = 'ouest',
        NORD_OUEST = 'nord_ouest',
        SUD_OUEST = 'sud_ouest',
    }
}